+++ modules/grid/src/test/java/integration/RemoteWebDriverTest.java
@@ -1,13 +1,10 @@
 package integration;
 
-import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.WebDriverRunner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.remote.RemoteWebDriver;
 
-import javax.annotation.Nonnull;
 import java.io.File;
 import java.io.FileNotFoundException;
 
@@ -17,7 +14,6 @@ import static com.codeborne.selenide.Selenide.closeWebDriver;
 import static com.codeborne.selenide.files.FileFilters.withExtension;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
-import static org.openqa.selenium.remote.CapabilityType.ACCEPT_INSECURE_CERTS;
 
 public class RemoteWebDriverTest extends AbstractGridTest {
   private RemoteWebDriver driver;
@@ -30,22 +26,12 @@ public class RemoteWebDriverTest extends AbstractGridTest {
 
   @Test
   void canOpenCustomRemoteWebDriver() throws FileNotFoundException {
-    driver = new RemoteWebDriver(gridUrl, getChromeOptions());
+    driver = new RemoteWebDriver(gridUrl, chromeOptions(null));
     WebDriverRunner.setWebDriver(driver);
     openFile("page_with_uploads.html");
 
     File downloadedFile = $(byText("Download a PDF")).download(withExtension("pdf"));
 
     assertThat(downloadedFile.getName()).isEqualTo("minimal.pdf");
-  }
-
-  @Nonnull
-  private static ChromeOptions getChromeOptions() {
-    ChromeOptions options = new ChromeOptions();
-    options.setCapability(ACCEPT_INSECURE_CERTS, true);
-    if (Configuration.headless) {
-      options.addArguments("--headless=new");
-    }
-    return options;
   }
 }