+++ src/test/java/com/codeborne/selenide/webdriver/EdgeDriverFactoryTest.java
@@ -54,6 +54,7 @@ class EdgeDriverFactoryTest {
       "--proxy-bypass-list=<-loopback>",
       "--disable-dev-shm-usage",
       "--no-sandbox",
+      "--remote-allow-origins=*",
       "--window-size=1366,768"
     );