+++ statics/src/main/java/com/codeborne/selenide/WebDriverRunner.java
@@ -187,25 +187,12 @@ public class WebDriverRunner {
     return webdriverContainer.hasWebDriverStarted();
   }
 
-  public static void using(WebDriver driver, Runnable lambda) {
-    if (hasWebDriverStarted()) {
-      var previous = getWebDriver();
-      try {
-        lambda.run();
-      }
-      finally {
-        setWebDriver(previous);
-      }
-    }
-    else {
-      setWebDriver(driver);
-      try {
-        lambda.run();
-      }
-      finally {
-        webdriverContainer.resetWebDriver();
-      }
-    }
+  static void using(WebDriver driver, Runnable lambda) {
+    webdriverContainer.using(driver, null, lambda);
+  }
+
+  static void using(WebDriver driver, SelenideProxyServer proxy, Runnable lambda) {
+    webdriverContainer.using(driver, proxy, lambda);
   }
 
   @CheckReturnValue