+++ statics/src/test/java/integration/SelenideMethodsTest.java
@@ -23,12 +23,14 @@ import static com.codeborne.selenide.Condition.cssValue;
 import static com.codeborne.selenide.Condition.disabled;
 import static com.codeborne.selenide.Condition.disappear;
 import static com.codeborne.selenide.Condition.empty;
+import static com.codeborne.selenide.Condition.exactOwnText;
 import static com.codeborne.selenide.Condition.exactText;
 import static com.codeborne.selenide.Condition.exist;
 import static com.codeborne.selenide.Condition.focused;
 import static com.codeborne.selenide.Condition.have;
 import static com.codeborne.selenide.Condition.hidden;
 import static com.codeborne.selenide.Condition.id;
+import static com.codeborne.selenide.Condition.innerText;
 import static com.codeborne.selenide.Condition.name;
 import static com.codeborne.selenide.Condition.partialText;
 import static com.codeborne.selenide.Condition.readonly;
@@ -169,6 +171,17 @@ final class SelenideMethodsTest extends IntegrationTest {
 
     assertThat($("#theHiddenElement").innerText().trim())
       .isEqualTo("Видишь суслика? И я не вижу. А он есть!");
+
+    $("#theHiddenElement").shouldHave(exactOwnText("Видишь суслика? И я не вижу. !"));
+    $("#theHiddenElement").shouldHave(innerText("Видишь суслика? И я не вижу. А он есть!"));
+  }
+
+  @Test
+  void innerText_errorMessage() {
+    assertThatThrownBy(() -> $("#theHiddenElement").shouldHave(innerText("Видишь суслика?")))
+      .isInstanceOf(ElementShould.class)
+      .hasMessageStartingWith("Element should have inner text \"Видишь суслика?\" {#theHiddenElement}")
+      .hasMessageContaining("Actual value: text=\"Видишь суслика? И я не вижу. А он есть!\"");
   }
 
   @Test